Move more code into mbc.cpp, and make `fatal()` actually exit
