+++ test/fix/noexist.err
@@ -1,2 +1,2 @@
-FATAL: Failed to open "noexist" for reading+writing: No such file or directory
+error: Failed to open "noexist" for reading+writing: No such file or directory
 Fixing "noexist" failed with 1 error
