+++ test/fix/tpp1-bad-major.err
@@ -1,1 +1,1 @@
-error: Failed to parse TPP1 major revision number
+FATAL: Failed to parse TPP1 major revision number
