+++ test/fix/tpp1-bad-minor.err
@@ -1,1 +1,1 @@
-error: Failed to parse TPP1 minor revision number
+FATAL: Failed to parse TPP1 minor revision number
