+++ test/fix/tpp1-too-short.err
@@ -1,2 +1,2 @@
-FATAL: "<filename>" too short, expected at least 340 ($154) bytes, got only 339
+error: "<filename>" too short, expected at least 340 ($154) bytes, got only 339
 Fixing "<filename>" failed with 1 error
