+++ test/fix/tpp1-unk-major.err
@@ -1,1 +1,1 @@
-error: RGBFIX only supports TPP1 version 1.0
+FATAL: RGBFIX only supports TPP1 version 1.0
