+++ test/fix/tpp1-unk-minor.err
@@ -1,1 +1,1 @@
-error: TPP1 minor revision number must be 8-bit
+FATAL: TPP1 minor revision number must be 8-bit
